requirements.txt
youtube-dl>=2019.7.12
PySide>=5.12.4

